Removed methods that aren't actually supported by selenium remote webdriver server
The endpoints:
/session/:sessionId/window/:windowHandle/size
/session/:sessionId/window/:windowHandle/position
/session/:sessionId/window/:windowHandle/maximize

Don't actually support dynamic :windowHandle as parameter, it always refers to the current focused window instance, no matter what's being passed it.
What this means is that getting/setting a window's size and position and maximizing is limited only to current window. This commit removes the methods that imply that other windows different than the current one can be manipulated in such manner.
Only size, position and maximizing dealings are available to current focused window.